Make integrations as separate entity, which might be also and entity of guild
